requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
gunicorn==22.0.0
google-generativeai==0.8.3
firebase-admin==6.5.0cryptography
